requirements.txt
python-telegram-bot==20.7
opencv-python==4.8.1.78
pyzbar==0.1.9
Pillow==10.1.0
numpy==1.24.4
python-dotenv==1.0.0